recovery: Remove deprecated WifiRouterHost.board usage
The WifiRouterHost.board value (also known as build_target) was
previously deprecated and was not thought to be used anymore. There
was one usage still remaining in the Gale provisioning version
check though, which is now failing because this value is no longer set.
This removes the unnecessary check of "board", as checking the "model"
is sufficient to check that the router is a Gale. Also, since logic is
dependent on this string value, it has been switched to be a constant
to ensure consistency.

BUG=b:297034827
TEST=cd ~/cr/infra/go/src/infra/cros/recovery \
&& go clean && go build && go test ./...

Change-Id: Ia9d0b0e5484fc0908068341321768f48e0bfa93f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4814566
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#59248}