config: Set model before Gale version check
The Gale version check requires model to be set on the WifiRouterHost
or it will fail, but the model was not set until after this check.
This worked fine for hosts that already had model set, but new
shivas deployments do not start with this value set and fail as a
result. This simply reorders the Gale repair steps to set the model
before checking the version so that it is always set.

BUG=b:297034827
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && \
go clean && go build && make tree && go test ./...

Change-Id: I4d0e2e9621bae525cc1596fc96139b77b115202a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4814888
Auto-Submit: Jared Bennett <jaredbennett@google.com>
Commit-Queue: Otabek Kasimov <otabek@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#59262}