[golangbuild] disable the dry-run-based cherryPick flag for sourceSpecs
Currently we set cherryPick to true on the buildSpec's sourceSpec if
it's a CV dry-run, but because we don't pass CV's properties to child
builds, we end up cherry-picking when we didn't mean to (i.e. on a
dry-run).

We could start doing that, but it would require a change to luci-go. As
a stop-gap, just act as if we only ever get dry-runs. We never actually
exercise the non-dry-run path intentionally at the moment. Just leave a
TODO explaining the situation.

Change-Id: I77e2c2be0600d3446a18bd181b50ed1719716999
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4823805
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#59347}