wifi: Add -n flag to OpenWrt sysupgrade cmd
We never want to preserve configs during a sysupgrade, as our images
contain our golden configs for OpenWrt devices. This adds the "-n"
flag to the sysupgrade call to not keep any config settings from the
current installed image so that it does not interfere with the new
images' configs.

BUG=b:296861897
TEST=manually preformed sysupgrade with -n flag with our test aps
successfully
TEST=cd ~/cr/infra/go/src/infra/cros/recovery \
&& go clean && go build && go test ./...
TEST=successfully ran OpenWrt repair locally with upgrade step ran

Change-Id: I458d2cad0ce3146db41048ac432ed3ac4b206d3f