wifi: Reboot and retry if OpenWrt update fails
We have an ongoing issue where there is an unknown process that
is wiping the /tmp dir on OpenWrt hosts, which cause them to break
in a bunch of different places, including during sysupgrade. This
adds another recovery step to reboot and then retry the update
if the first attempt fails, which will allow repair to recover these
routers.

BUG=b:296861897
TEST=successfully ran router recovery with OpenWrt hosts with and
without a wiped /tmp dir where the normal upgrade action and the new
upgrade actions worked as expected to recover the devices
TEST=cd ~/cr/infra/go/src/infra/cros/recovery \
&& go clean && go build && make trees && go test ./...
TEST=cd ~/cr/infra/go/src/infra/cmd/mallet && go clean && go build

Change-Id: I28f60e4b2dec85a59a4fb70673231ce5e72c0d6d