config: start using factory CCD setting for all Ti50 devices
With the release of the AP RO verification feature, all ti50 devices
regardless of if the main servo connection is through CCD or not, need
to have CCD open and reset to factory settings. This ensures that the
AllowUnverifiedRO CCD capability is set to always.

BUG=b:298237474
TEST=none

Change-Id: I7189ea88f6b4000af9dba054c1bea98c0bacb6d6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4834458
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jett Rink <jettrink@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#59453}