[test-health] Fallback to test_id when test_name not populated
About 1000 tests are missing a test name. We still want to know how
much resources this test is using so fall back to display the test_id
when test_name is missing. This would also make a confusing experience
if a user tries to filter for these so we need to include the test_id
in the filter now as well.

Tests with any missing data would also not appear when filtered because
a null concatenated with a string will return null. So replace
potentially null entries with an empty string for filtering

Bug: 1435578
Change-Id: I194c618f13ec1bfe9117aa4f5daf33e12d04fd6e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4847259
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Riley Wong <rgw@google.com>
Cr-Commit-Position: refs/heads/main@{#59534}