[test-health] Fix time display bug
For seconds less than 0, we want to surface up to 2 precision points if
possible to denote that there was a run time as 0 seconds is misleading.

For anything less that 0.01 seconds, we will denote as <0.01s and we will
truncate the decimal if mins/hours/day are present.

Bug: 1435094
Change-Id: Ic82ea18bd4449b1487bdd48360110eda3e83689e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4852388
Commit-Queue: Gary Tong <gatong@chromium.org>
Reviewed-by: Gary Tong <gatong@chromium.org>
Auto-Submit: Riley Wong <rgw@google.com>
Cr-Commit-Position: refs/heads/main@{#59639}