[golangbuild] distribute ports across misc-compile shards more evenly
When the sharded version of the misc-compile builder runs at the tip of
the main Go repo, there are 35 ports that are sharded across 12 shards.

The balance is fair but not great: two of the shards get 5 ports,
some get 1 port. It's easy to make the sharding much better by
assigning ports to evenly-ish-divided buckets (3+3+3+...+3+2 = 35).

Change-Id: I080d2758bd8e85795c30afaac95fe057aa9858b8