[golangbuild] rebalance distibution of tip ports across 12 shards
When the sharded version of the misc-compile builder runs at the tip of
the main Go repo, there are 35 ports that are sharded across 12 shards.

The balance is fair but not great: two of the shards get 5 ports,
some get 1 port. It's easy to make the sharding much better by finding
a good sharding salt. Do that, along with a manual test so that it's
easier to run this code again to rebalance when occasionally needed.

Change-Id: I080d2758bd8e85795c30afaac95fe057aa9858b8