recovery: Add CurlFileContents cache util
Adds a new cache utility function, CurlFileContents, that is the same
as the existing CurlFile function except that instead of saving the
file to the host, it collects and returns the file contents as a
string. The curl command used differs in that no output file is
specified, so it is printed to stdout and no file is created on the
host system for this purpose. This eliminates the need for a temporary
file on the host when the file contents is expected to be normal text.

BUG=b:291670455
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && \
go clean && go build && go test ./...
TEST=successfully used in later CLs to read JSON files from the cache

Change-Id: I5723b16b085ec1e253edcb29d9f9d6268c0385ca