[golangbuild] pass exact GoCommit to triggered builds when invoked != go
The goSrc field in buildSpec holds the exact Go repo spec to fetch and
use for a given build. When invoked in "go" repo, its value comes from
invokedSrc, which in the sharded builder case will be the same for all
of coordinator, build-go, and test-only triggered builds.

When invoked in an x repo, the tip of inputs.GoBranch is read. This is
fine to do in the coordinator, but since a branch is a moving target, it
needs to be passed from coordinator to builds it triggers. Add and use
an optional GoCommit input property for that.

Change-Id: I1039eb65a380017b5b08b6734e22db36ab064466
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4864775
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#59780}