recovery: Add btpeer ssh runner
Starts a new btpeer components package with a generic ssh.Runner
implementation meant to be used with btpeer hosts.

This is technically using the ssh utils that were built for wifi
router execs and still remain under the wifi package, but they are
generic utils not specific to wifi hosts. We have a TODO to refactor
these utils for general use, and this new usage will be refactored
at the same time.

This implementation is essentially the same as the wifi router one,
but without the ssh username override logic as all btpeers use the
normal default credentials for ssh access. When we refactor these
utils, this will likely be the default implementation.

BUG=b:291670455
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && \
go clean && go build && go test ./...

Change-Id: Idba7c55b5e4ecf27759df686acf932d0d0d32bf1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4867470
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Cr-Commit-Position: refs/heads/main@{#59955}