[ufs] Defensively add _extra field to every entity.
When handling a recent P0, we rolled back UFS and that resulted in some unreadable records.

There are two potential sources of unreadable records in the UFS code: one is through the datastore ORM. And the other is unmarshaling base64 (or however they're encoded) protos in UFS.

This CL addresses one of those potential sources of unreadable records.

BUG=b:302015191
TEST=make test

Change-Id: I7c89ad5aa66c4d43bc54124c1379a214b1a106cb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4892664
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#59992}