recovery: Add btpeer chameleond release selection
This copies the go implementation and unit tests of the btpeer
chameleond release selection logic from the btpeer_manager CLI
util into infra so that it may be used in repair execs. The only
changes are the libraries used for error reporting and logging.

This logic is defined in the comments of the
BluetoothPeerChameleondConfig proto message that these functions
use, and is currently in use in production today in tauto tests
that update btpeers.

BUG=b:291670455
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && \
go clean && go build && go test ./...

Change-Id: Iecd73e6b582d1389f6df885c0b2dd44071def728
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4867471
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#59997}