Manual Reland of [ufs] Defensively add Extra field to every entity.
When handling a recent P0, we rolled back UFS and that resulted in some unreadable records.

There are two potential sources of unreadable records in the UFS code: one is through the datastore ORM. And the other is unmarshaling base64 (or however they're encoded) protos in UFS.

This CL addresses one of those potential sources of unreadable records.

BUG=b:302015191
TEST=make test

Change-Id: I94527bf692184c62b5dc5b09ad46a1225ee890e5