recovery: Set peripheral_btpeer_state
Adds support for setting the new peripheral_btpeers_state
field based on the states of all btpeers. Since this is just
based on other data we already collect during repair, we only
need to update this one place in repair to set this field.

BUG=b:302033368
TEST=cd ~/cr/infra/go/src/infra/cros/recovery \
&& go clean && go build && make gen && go test ./...

Change-Id: Iac742aa37e0b6ad660ff41384df1c5a6d05917d6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4893326
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#60114}