recovery: Update btpeer config to update chameleond
Adds new set of actions to the btpeer repair config that will
use new btpeer execs to check the chameleond version on the
btpeer against the expected release and updates it if it does
not match. A new reboot action and check to make sure chameleond
is running were added as well to keep and validate the health
of the device.

Old actions that were not in use (because they didn't work) which
checked if chameleond was responsive have been removed in favor
of these new actions.

BUG=b:291670455
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && \
go clean && go build && go test ./...
TEST=successfully ran btpeer repair locally with multiple testbeds

Change-Id: I91486f39df78e46deb7fada68b5a250b930fed33
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4867473
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Cr-Commit-Position: refs/heads/main@{#60126}