recovery: Remove unused btpeer exec
Removes the unused getDetectedStatusesExec btpeer exec and the utility
it used to make chameleond calls. This exec has been disabled for a
long time since the chameleond call util did not work in the lab, and
with the last config update the actions that used this exec have been
deleted.

If we need to ever make chameleond calls during infra repair (which
at this time we do not plan to need to) we can copy over the much more
feature-rich golang chameleond XMLRPC API we made for tast and/or find
a way to share it between the two repos.

BUG=b:291670455
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && \
go clean && go build && go test ./...

Change-Id: I2a69a7604d1695a20f36c1d94c4e5e49804999f4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4867474
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#60127}