siso: fix flush symlink
When prepare inputs, flush is called for inputs.
In ChromeOS simplechrome flow, inputs contains
symlinks for toolchain/sysroots.

Hashfs Flush tried os.Symlink and if exists, remove and create again,
but it may be racy.
Check existing symlink first and do nothing if target is the same.
If err or target differs, create/recreate symlinks.

Bug: b/303177370
Change-Id: I4b656de3abf5d5b9e0341326cf52692f4e12b03a