Check for presence of PresubmitCoverageData entity before blocking.
Blocking a CL for Low coverage works in the following way: when we process a pubsub request for a coverage build, we mark that build "seen". When the last such build is marked as "seen", we check if the CL can be blocked for low coverage. This tracking works happen through a LowCoverageBlocking entity.

Sometimes, none of the coverage builds generate coverage data, resulting in no PresubmitCoverageData entity. But when the last build gets marked "seen", we check for blocking low coverage and it fails because no corresponding PresubmitCoverageData entity exists, resulting in a 500 error

Adding a check for presence of PresubmitCoverageData before checking for blocking would solve this.

Bug: 1491880
Change-Id: I7b9173f6107e65010005a765289bb24271165b04
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4932469
Commit-Queue: Prakhar Asthana <pasthana@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Auto-Submit: Prakhar Asthana <pasthana@google.com>
Cr-Commit-Position: refs/heads/main@{#60338}