[golangbuild] surface all source input information to UI and properties
Currently, both in the UI and programmatically, it's difficult to figure
out exactly what commits went into a particular subrepo build. The root
cause of the problem is that a build can only be triggered by one
source, so we pull HEAD of the right branch for either the subrepo or
main Go repo, whichever one we don't have. However, we don't put this
commit information anywhere.

This change puts the source information into the output properties and
also places it in the summary markdown for the build.

While we're here, this change also adds an explicit
SUCCESS/FAILURE/INFRA FAILURE message at the top of the summary
markdown, something that currently gets lost when we set the summary
markdown.

Change-Id: If32288a9498a56cae07d79fe856d69a15f807907
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4929223
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#60341}