[experimental] Remove __init__.py
I want to put some go code in here and this experimental directory
being a Python module slightly gets in the way of that.

This isn't a real obstacle, but, like, the hygiene is bad and this
seems like as good an opportunity as any to take care of that.

BUG=None
TEST=None

Change-Id: I36be279751924dce84f2e91613f28075ee49af78
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4937775
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#60383}