recovery: rename "Is not Flex device" action in servo
The repair action "Is not Flex device" no longer applies just to reven (see crrev/c/4521582). This is part of a series of changes that adjust the action name to be more inclusive as well as updating the doc strings to better reflect what it does.

BUG=b:284342200
TEST=cd ~/infra/go/src/infra/cros/recovery && go clean && go build && make trees && go test ./...

Change-Id: I7e08a77471d76fa6a064da74cc79a1d68f295d44
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4950256
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Andre Braga <andrebraga@google.com>
Cr-Commit-Position: refs/heads/main@{#60475}