recovery: rename "Is not Flex device" action in closing
The repair action "Is not Flex device" no longer applies just to reven (see crrev/c/4521582). This is part of a series of changes that adjust the action name to be more inclusive as well as updating the doc strings to better reflect what it does.

BUG=b:284342200
TEST=cd ~/infra/go/src/infra/cros/recovery && go clean && go build && make trees && go test ./...

Change-Id: I1e72de018b80ce209ccdf68806f81e4bc9fc95ee
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4950257
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Andre Braga <andrebraga@google.com>
Cr-Commit-Position: refs/heads/main@{#60476}