satlab: Restore add DUT cmd output
Previously we attempted to refactor the add DUT common library to remove
print statements (as printing info to stdout may not always be
appropriate for a common library).

This CL restores the output but accepts an io.Writer as input for use
cases where printing to stdout is less desirable. It also changes return
types to be a bit more generic so that a future replacement with UFS api
will be more natural.

BUG=b/306444923
Test Plan=make satlab and testing various flows

Change-Id: I588593994bc12256f7544679ad999d7764246bf1