[test-health] Add hyperlink to test suites that cause overflow in directory view
- Certain test files have more test variants than we can show within the
directory view dropdown. This causes the sum of metrics not be equal to
the test variants shown and also misleads the user into thinking there
are no more test variants.
- To mitigate this issue, we will provide a hyperlink to a different
page with the test file name as the filter. This will surface all the
test variants related to that test file.
- This impacts directory view in snapshot view mode and time line view mode.
- This introduces table footers to our makeshift table. Developers
will be able to customize footers from now on.

Bug: 1435094
Change-Id: If1859f950ff520ef967033e642dd75dccbb8f3dc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4936151
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Riley Wong <rgw@google.com>
Cr-Commit-Position: refs/heads/main@{#60587}