satlab: Restore add DUT cmd output
Previously we attempted to refactor the add DUT common library to remove
print statements (as printing info to stdout may not always be
appropriate for a common library).

This CL restores the output but accepts an io.Writer as input for use
cases where printing to stdout is less desirable. It also changes return
types to be a bit more generic so that a future replacement with UFS api
will be more natural.

It's likely this doesn't replicate the EXACT output of the command but
it should provide an appropriate amount of information to users.

BUG=b/306444923
Test Plan=make satlab and testing various flows

Change-Id: I588593994bc12256f7544679ad999d7764246bf1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4955808
Reviewed-by: Prasad Vuppalapu <prasadv@chromium.org>
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Cr-Commit-Position: refs/heads/main@{#60593}