cros_test_runner: Remove direct references to dut topology
Some commands are still calling directly into the dut topology to grab the primary dut when that information has already been parsed and stored in the state keeper. Ignoring the parsing has the side effect of referencing the incorrect dut as the primary based on the request details.

Remove certain direct calls to the dut topology. Opting to not remove references within VM area as I am not as certain when making changes there.

BUG=b:306679104
TEST=led

Change-Id: I6411cf4831bb8c20dc5e28e8b1c5d3e0f6093e81
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4969872
Auto-Submit: Chris DeLaGarza <cdelagarza@google.com>
Reviewed-by: Azizur Rahman <azrahman@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Cr-Commit-Position: refs/heads/main@{#60604}