shivas: Include peripheral labels in su labels
There are peripheral-related dimensions within dut dimensions, but
are not present in the scheduling unit dimensions. This adds versions
of these peripheral-related dims to scheduling unit's dims based on
the combination of all duts in the unit. This allows for SUs to be
scheduled against these peripheral-related dims like testbeds with
just one dut can be.

BUG=b:302749229
TEST=cd ~/cr/infra/go/src/infra/cmd/shivas && make
TEST=cd ~/cr/infra/go/src/infra && golangci-lint run \
cmd/shivas/utils/schedulingunit/...

Change-Id: I6ecef8dd5f64a32ec1dbc87191aadb9b5efd05b6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4977065
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#60688}