[mac_toolchain] specify non-interactive sudo
Currently if a bot is misconfigured to not allow passwordless sudo for
these commands it will silently hang waiting for password input, which
is painful to debug.

By specifying non-interactive sudo, if a password is required sudo will
fail fast with "sudo: a password is required".

Change-Id: I81bf8bed7bea55e1b777f3c0e76929d9b4290782
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4985122
Reviewed-by: Yue She <yueshe@google.com>
Commit-Queue: Michael Pratt <mpratt@google.com>
Cr-Commit-Position: refs/heads/main@{#60731}