recovery: Use dutName for peripheral GetCacheUrl calls
The wifirouter and btpeer repair execs that used GetCacheUrl are
passing their resource name, but this needs to be the dut hostname.
Since there is somewhat support for this as it attempts to extract dut
hostname from the passed name, this is working fine in most cases.
However, when repair is run on multidut testbeds whose dut hostnames
differ from their usual format, the dut hostname extraction results in
an incorrect dut hostname, causing GetCacheUrl to fail.

To fix this issue, all usage of GetCacheUrl by wifirouter and btpeer
repair execs have been updated to pass the dut hostname directly. Also,
the GetCacheUrl function argument name has been refactored from
"resource" to "dutName" to make it more clear that it is supposed to be
the dut name and not just any resource name.

BUG=b:307915695,b:302749229
TEST=successfully ran wifirouter and btpeer repair plans on both a
testbed with the regular dut hostname format and with the irregular
hostname format and validated that the cache access was successful via
the logs
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && go clean && go build \
&& make trees && go test ./...

Change-Id: I0c3fb089c56e01cd9a3271c57043deed22721708
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4980817
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Cr-Commit-Position: refs/heads/main@{#60739}