config: Add metrics dir to gale stateful cleanup
A metrics dir is created for each shutdown of a gale, and it is
stored in the stateful partition that persists. Thus, as more
shutdowns occur, more metrics logs are created. Over many reboots,
this can cause the stateful partition to fill up beyond the 50%
threshold. Since we don't clean these up, the step meant to clean
the stateful partition can fail to clean it up to under 50%. This
is more common now that we are always rebooting the gale router on
repair. This adds the metrics dir to the cleaned dirs of the
stateful partition cleanup step to prevent this scenario.

BUG=b:302225472
TEST=cd ~/cr/infra/go/src/infra/cros/recovery && \
go clean && go build && make trees && go test ./...
TEST=successfully ran wifirouter repair plan locally and confirmed
the metrics were deleted on a router exceeding the stateful partition
size threshold

Change-Id: I842a134641ebe757fae0ba9f17b631f7d0050aa5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4985759
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Cr-Commit-Position: refs/heads/main@{#60764}