config: Do not provision Gales after cleanup failure
When a Gale drive cleanup fails, the next recovery step is to
provision the Gale. Recent repair failures show that this only makes
the issue worse, as provisioning when there isn't enough drive space
will cause the provisioning to fail before it completes, resulting in
an inaccessible Gale that can only be fixed via manual flashing. So,
this removes that step. It's better to have it just fail at the
cleanup, especially with the upcoming logs that will allow us to
investigate and fix drive cleanup issues without needing manual
repairs.

BUG=b:302225472
TEST=cd ~/cr/infra/go/src/infra/cros/recovery \
&& go clean && go build && make trees && go test ./...

Change-Id: If8556d919c1e303c5aaa4752c3bccada97e01a89
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4989906
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Cr-Commit-Position: refs/heads/main@{#60765}