[ufs] In network.go make as many things unexported as possible.
It's too large of a code change to change the API of the util
functions that take strings to take net.IPs instead, so I am
going to change strategy. I will update all of the helper functions
to use net.IPs internally, and then, as the last step, extend the
interface of the public functions to accept IPv6 addresses.

BUG=b:187899200
TEST=make test

Change-Id: I29a4644ecb414bce9fdd4b96655b7ee0724f1334
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5005126
Reviewed-by: Anushruth Sasivehalli <anushruth@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#60923}