[ufs] Enforce limits on number of IP items created during vlan creation
We don't limit the overall size of the vlan, but this CL does impose a new limit.
We now no longer allow more than 2000 initial reserved IPs or final reserved IPs.

I think this change is for the best.

However, as of this CL, we will only create up to 2000 IP objects for each vlan.
This change is not observable. Trying to get a new IP address will preferentially grab
the preallocated ones, which is cool.

BUG=b:187899200
TEST=make test

Change-Id: I539f8bbd4daf4656862af9ada0f896cf93fdc33e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4987672
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Yang Yang <echoyang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#60957}