[test-health] Correct average cores based on real data
Rdb durations are created by each test harness so the time can be
measured differently depending on the builder/test suite. avg_cores is
a metric focused on the actual resource consumption of each test. To
keep these variances in harnesses from affecting the metric this will
measure actual swarming tasks and the sum total of rdb tests that
came from that task. The swarming time used divided by the rdb
reported run time will give us the amount of time the test is
responsible for the swarming task to complete, giving us a normalized
resource usage of "bots". The cores argument present in the swarming
tasks is also multiplied by the swarming duration to get the total
available computational time which can then be used to get the
average "cores". The current time is also used to normalize on time
which it should be noted has little to do with the peak usage and won't
map 1 to 1 to the allotted resources for the test

Bug: 1435578
Change-Id: I619693fe4ea9c3d69f0257a81e263d5a068a5f98
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4906189
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Gary Tong <gatong@chromium.org>
Reviewed-by: Stephanie Kim <kimstephanie@google.com>
Cr-Commit-Position: refs/heads/main@{#61093}