suitescheduler: build CTP request for a given susch config
One thing also in this change is an update to the SuSch ingestion
system. This change fixes the logic that was used in the past to properly generate all board(variant)/model combos according to the SuSch rules. This was included in the CL because it was a requirement for properly generating the CTP Requests.

BUG=b:305740471
TEST=https://chromeos-swarming.appspot.com/task?id=65bf847779131c10

Change-Id: Icd4d2f8f623119a809fcbdd8272d767ef5a4b2ec
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4953492
Commit-Queue: Julio Hurtado <juahurta@google.com>
Reviewed-by: Azizur Rahman <azrahman@google.com>
Cr-Commit-Position: refs/heads/main@{#61247}