golangbuild: support changes outside of go.googlesource.com
Relui is starting to use buildbucket for testing releases, which means
we have to be able to run tests off commits in the private repo. Attempt
to support that.

Despite the TODO, it *looks* like the primary problem is just that we
don't recognize golang/go-private as being the main repo. Treat it as
such. It might have been possible to rewrite the project name to "go",
but it seemed way too confusing to track which was the physical vs.
logical project name. This will be easy enough to unwind when we move
to go-internal.googlesource.com.

For golang/go#63147.

Change-Id: If4c3937306b48e32d4e64972257255aebf680d12
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5027587
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Heschi Kreinick <heschi@google.com>
Cr-Commit-Position: refs/heads/main@{#61249}