[ufs] Enforce validation of entities when writing to datastore.
Add a layer of validation to UFS right before writing data to datastore.
Some data, including IPEntities which I have recently modified, have
well-formedness conditions that are somewhat complex. (For example, the name
of an IPEntity must match the IPv4 address (encoded using the traditional
127.0.0.1 notation) or the IPv6 address (encoded using hex).

I think we should change UFS to validate data when we write it, but allow
invalid data to be read.

This has advantages and disadvantages. The validation logic might be buggy in
practice. It might cause problems during a rollback.

However, it also has the potential to prevent bad data from being written in
the first place, which I think is compelling.

Change-Id: I9dcc65b5f8e1473cae3f092fe2fabf248cc63185
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5027689
Reviewed-by: Yang Yang <echoyang@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#61255}