Add -start-step-tag and -end-step-tag args to compilator_watcher
In preparation for the compilator_watcher using -start-step-tag
and -end-step-tag flags instead of -swarming-trigger-phase and
-local-test-phase, this change allows the orchestrator to send in
both the old flags and these new flags. Then, the compilator_watcher
will be updated to use the new flags.

The `_ = <variableName>` is there to ignore errors about the variable
being unused.

Bug: 1248460
Change-Id: Id98beaf30ed4766e8ebc4e86c94a05f23086b08f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5031408
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Stephanie Kim <kimstephanie@google.com>
Cr-Commit-Position: refs/heads/main@{#61299}