[fleetcost] Create do-nothing server
Create a do-nothing server that can run locally and define a Dockerfile. Cloud
run needs a Dockerfile.

Also, change the name of the executable to "fleetcostserver" so that the name
"fleetcost" is available for a command line tool.

In my experience, developping a service like this is much easier with a local
command line tool that can perform administrative tasks, or just hit RPC
endpoints.

BUG=b:311443080
TEST=make test

Change-Id: I687b75ce71034a59bdc4eac3a21ec37f02857feb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5038326
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#61311}