siso: add cleandead
Add cleandead to clean built files that are no longer produced by the manifest
in the same way as ninja.

It checks generated files recorded ninja log and check it has a node
in build graph and it has in edge or out edges.

In Siso, generated files are recorded in hashfs, which is associated
with cmdhash.
Siso recorded generated files including depfile, but depfile is not
a node in build graph (and ninja also doesn't record depfile in ninja log),
so changed not record depfile with cmdhash.

Bug: b/312557439
Change-Id: Ie1dbe75450b8b623e4fdc76c3e7e56e5b9bf5e28
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5051744
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#61456}