🚝 Fix issue submission with issue wizard with Chrome 121 and up.
The issue wizard attempts to add the Needs-Triage-M### label with the
current user agent's Chrome version. This label exists for M120 and below. Since label creation is frozen in Monorail, this is a 400 error
for Chrome 121 and up.

So don't add the label for Chrome 121 and up.

Bug: b/312712149
Change-Id: I9bb046f27a336a531c22861b2fb959f5622b1b77
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5073932
Commit-Queue: Dave Tu <dtu@chromium.org>
Reviewed-by: Kevin Ruvalcaba <kevinruvalcaba@google.com>
Auto-Submit: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#61548}