golangbuild: also set GOHOSTOS/GOHOSTARCH globally, run go env
It seems that tests misbehave if the GOHOST vars don't match the
toolchain builds -- see the rollback at go.dev/cl/545157.

I haven't been able to reproduce this, so it's just a hypothesis,
but easy to test.

In case it's not the issue, run `go env` to aid further debugging.

Change-Id: I1ddebcf16dc623aadc42ebb0cc6131f7fe779746
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5073933
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Heschi Kreinick <heschi@google.com>
Cr-Commit-Position: refs/heads/main@{#61554}