Remove dheriot@google.com from OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/dheriot@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I3d7e41266a2c51f9cec3541c31a1d029b8967546
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5074536
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Michal Matyjek <mmatyjek@google.com>
Commit-Queue: Michal Matyjek <mmatyjek@google.com>
Cr-Commit-Position: refs/heads/main@{#61583}