Remove mmatyjek@google.com from go/src/infra/cros/satlab/OWNERS
This suggested change is automatically generated based on commit
and review activity to improve the reliability of OWNERS suggestions
and CL review latency.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.
If no no CR -1 vote is set after 5 days, the CL will be submitted.
OWNERS expectations for the chromium project are at:
https://chromium.googlesource.com/chromium/src/+/refs/heads/main/docs/code_reviews.md#expectations-of-owners

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/mmatyjek@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I480a3aeb982de5b0d3f102f9f3a905b6e73e7641
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5076331
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Varun Thammineni <varunth@google.com>
Reviewed-by: Varun Thammineni <varunth@google.com>
Cr-Commit-Position: refs/heads/main@{#61605}