Remove ayatane@chromium.org from go/src/infra/cros/recovery/OWNERS
This suggested change is automatically generated based on commit
and review activity to improve the reliability of OWNERS suggestions
and CL review latency.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.
If no no CR -1 vote is set after 5 days, the CL will be submitted.
OWNERS expectations for the chromium project are at:
https://chromium.googlesource.com/chromium/src/+/refs/heads/main/docs/code_reviews.md#expectations-of-owners

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/ayatane@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ic03479312bbb7820f3ee32c7423c920e9a564879
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5076598
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Sergey Fetisov <sfetisov@google.com>
Reviewed-by: Sergey Fetisov <sfetisov@google.com>
Cr-Commit-Position: refs/heads/main@{#61632}