Remove otabek@chromium.org from go/src/infra/cros/satlab/satlab/OWNERS
This suggested change is automatically generated based on commit
and review activity to improve the reliability of OWNERS suggestions
and CL review latency.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.
If no no CR -1 vote is set after 5 days, the CL will be submitted.
OWNERS expectations for the chromium project are at:
https://chromium.googlesource.com/chromium/src/+/refs/heads/main/docs/code_reviews.md#expectations-of-owners

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/otabek@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Id8ce22c982b50894486981764edc03a11d301cec
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5080469
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#61633}