New algorithm for calculating builder health and low-value indicators
The current algorithm for determining a builder's health score based on
metrics from the preceding seven days has a drawback. After an unhealthy
builder is fixed, its owner must wait several days for the green healthy
indicator to appear.

This CL introduces a new approach to calculating the indicator. Instead
of relying on a single seven-day average, we will calculate the health
score for each of the previous seven days. The final health score will
be the highest of these individual scores. Consequently, if a builder is
fixed yesterday, it will become healthy today.

Change-Id: Ia1f50884439af7bf5da3985ff50c76532cab3ac4
Bug: 1503192
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5077378
Commit-Queue: Trung Nguyen <tne@google.com>
Reviewed-by: Gregory Guterman <guterman@google.com>
Cr-Commit-Position: refs/heads/main@{#61693}