Add migrated_id to Monorail V3 API GetIssue
This follows the method used by https://chromium-review.googlesource.com/c/infra/infra/+/4621419 to add to the V0 API. It only adds it to the GetIssue method.

The diff is big and scary. These are the instructions I followed:
```
* https://chromium.googlesource.com/infra/infra/+/refs/heads/main/doc/source.md#checkout-code
* eval `../../go/env.py `
* make prpc_proto_v3
```
It looks like the versions of protoc-gen-go and protoc changed causing this diff?

Did not add a test because hopefully this is a smallish change and Monorail will go away one day.

Change-Id: Ibd45320e4752d91431b865de9ba83218e65d5264
Bug: b/311422959
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5089870
Reviewed-by: Yuxuan Chen <yuxuanch@google.com>
Commit-Queue: Ravi Mistry <rmistry@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/main@{#61710}