[chrome-test-health] Add function to fetch cov reports for past year
The getCoverageReportsForLastYear function fetches the absolute code coverage reports for the last 365 days. These reports are specific to builder configuration which is supplied to this function as builder and bucket.

This is a helper function that'll be used in the API to pull the one commit hash per day for the last year.

Bug: 1471659
Change-Id: Ia06f1141606b046d1b0932b83deaad4df58aff16
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5089522
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Vidit Chitkara <viditchitkara@google.com>
Cr-Commit-Position: refs/heads/main@{#61732}