[dirmd] Monorail migration command for Monorail->Buganizer migration
This change introduces a new subcommand (monorail) to migrate Monorail
component metadata to Buganizer.

This requires a .textproto mapping file be passed in to determine
the mappings between Monorail and Buganizer. Only chromium project
metadata are migrated.

Files that specify a Monorail component that does not map to anything
will print to console. There are a handful of Monorail components that
no longer exist in Monorail, which will be cleaned up in separate
efforts.

This command should be deprecated once the migration is complete.

Bug: 1505875
Change-Id: I688c2abf2182916847dca5b7697ef041ef15288e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5079176
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Cr-Commit-Position: refs/heads/main@{#61838}