Remove warning advising not to use infra.git directly.
During the submodule migration it was recommended to checkout the
infra_superproject as a way to get the depenencies for the infra
repo. However this is no longer the case - the infra.git repo has
a fully self-contained set of DEPS now, and it is fine to check it
out by itself.

R=vadimsh

Change-Id: I60d0bb313aff0f8a6b565b3f49222e7e69e321d1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5118283
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#61858}