[crderiveinputs] Generate comprehensive inputs manifest from a chromium commit.
This tool is intentionally intertwined with the chromium checkout
process.

The intent is to get to a point where a full input manifest
can be generated from a chromium commit such that the input manifest
can be fed to a tool to fetch all the sources, and after that fetch
tool runs, the build can be executed offline.

R=ashishgandhi, bryner

Change-Id: I98bd59b94cccb81ca3125ce19ac3db9e65fe2493
Bug: 1511102
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5006319
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#61929}